requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
typer==0.12.5
numpy==2.0.1
sqlite-utils==3.36
PyYAML==6.0.2
requests==2.32.3

# Enhanced NeuralSync2 dependencies
xxhash==3.4.1
zstandard==0.22.0
psutil==5.9.5
aiofiles==23.2.1
asyncpg==0.29.0
redis==5.0.1
lz4==4.3.2
msgpack==1.0.7
orjson==3.9.10
uvloop==0.19.0
cython==3.0.5
